+++ src/opengeodeweb_viewer/object/schemas/apply_textures.json
