+++ src/opengeodeweb_viewer/rpc/mesh/protocols.py
@@ -11,7 +11,7 @@ from wslink import register as exportRpc
 
 # Local application imports
 from opengeodeweb_viewer.utils_functions import get_schemas_dict, validate_schema
-from opengeodeweb_viewer.object.methods import VtkObjectView
+from opengeodeweb_viewer.object.protocols import VtkObjectView
 
 
 schemas_dir = os.path.join(os.path.dirname(__file__), "schemas")
@@ -21,8 +21,8 @@ class VtkMeshView(VtkObjectView):
     def __init__(self):
         super().__init__()
 
-    @exportRpc(schemas_dict["toggle_object_visibility"]["rpc"])
-    def toggle_object_visibility(self, params):
+    @exportRpc(schemas_dict["set_visibility"]["rpc"])
+    def SetVisibility(self, params):
         print(schemas_dict["toggle_object_visibility"]["rpc"], flush=True)
         validate_schema(params, schemas_dict["toggle_object_visibility"])
         super().SetVisibility(params)
@@ -33,17 +33,17 @@ class VtkMeshView(VtkObjectView):
         validate_schema(params, schemas_dict["set_opacity"])
         super().SetOpacity(params)
 
-    @exportRpc(schemas_dict["toggle_edge_visibility"]["rpc"])
+    @exportRpc(schemas_dict["set_edge_visibility"]["rpc"])
     def setEdgeVisibility(self, params):
-        print(schemas_dict["toggle_edge_visibility"]["rpc"], flush=True)
-        validate_schema(params, schemas_dict["toggle_edge_visibility"])
+        print(schemas_dict["set_edge_visibility"]["rpc"], flush=True)
+        validate_schema(params, schemas_dict["set_edge_visibility"])
         print(f"{params=}", flush=True)
         super().SetEdgeVisibility(params)
 
-    @exportRpc(schemas_dict["toggle_point_visibility"]["rpc"])
+    @exportRpc(schemas_dict["set_point_visibility"]["rpc"])
     def setPointVisibility(self, params):
-        print(schemas_dict["toggle_point_visibility"]["rpc"], flush=True)
-        validate_schema(params, schemas_dict["toggle_point_visibility"])
+        print(schemas_dict["set_point_visibility"]["rpc"], flush=True)
+        validate_schema(params, schemas_dict["set_point_visibility"])
         super().SetVertexVisibility(params)
 
     @exportRpc(schemas_dict["set_point_size"]["rpc"])
