+++ src/opengeodeweb_viewer/rpc/mesh/schemas/set_color.json
@@ -1,5 +1,5 @@
 {
-  "rpc": "set_color",
+  "rpc": "mesh.set_color",
   "type": "object",
   "properties": {
     "id": {
