+++ src/opengeodeweb_viewer/rpc/mesh/schemas/set_point_size.json
@@ -1,5 +1,5 @@
 {
-  "rpc": "set_point_size",
+  "rpc": "mesh.set_point_size",
   "type": "object",
   "properties": {
     "id": {
