+++ src/opengeodeweb_viewer/rpc/mesh/schemas/set_point_visibility.json
@@ -1,5 +1,5 @@
 {
-  "rpc": "toggle_object_visibility",
+  "rpc": "mesh.set_point_visibility",
   "type": "object",
   "properties": {
     "id": {
