--- src/opengeodeweb_viewer/rpc/viewer/schemas/delete_object_pipeline.json
@@ -1,11 +1,0 @@
-{
-  "rpc": "delete_object_pipeline",
-  "type": "object",
-  "properties": {
-    "id": {
-      "type": "string"
-    }
-  },
-  "required": ["id"],
-  "additionalProperties": false
-}
