+++ src/tests/data/images/mesh/display_vertex_attribute_1.jpeg
