+++ src/tests/data/images/mesh/display_vertex_attribute_2.jpeg
