+++ src/tests/data/images/mesh/toggle_edge_visibility.jpeg
