+++ src/tests/data/images/model/set_model_components_visibility.jpeg
