+++ src/tests/data/images/model/set_model_mesh_visibility.jpeg
