+++ src/tests/data/images/object/apply_textures.jpeg
