+++ src/tests/data/images/object/deregister.jpeg
