+++ src/tests/data/images/object/set_color.jpeg
