+++ src/tests/data/images/object/set_opacity.jpeg
