+++ src/tests/data/images/object/set_point_size_1.jpeg
