+++ src/tests/data/images/object/toggle_object_visibility_1.jpeg
