+++ src/tests/data/images/object/toggle_object_visibility_2.jpeg
