+++ src/tests/data/images/object/toggle_object_visiblity.jpeg
