+++ src/tests/data/images/object/toggle_point_visibility.jpeg
