+++ src/tests/test_viewer_protocols.py
@@ -14,62 +14,13 @@ def test_set_viewer_background_color(server):
     server.call("set_viewer_background_color", [{"red": 0, "green": 0, "blue": 255}])
     assert server.compare_image(3, "viewer/set_viewer_background_color.jpeg") == True
 
-def test_create_object_pipeline(server):
-    server.call("create_object_pipeline", [{"id": "123456", "file_name": "hat.vtp"}])
-    assert server.compare_image(3, "viewer/create_object_pipeline.jpeg") == True
-
-
-def test_delete_object_pipeline(server):
-    server.call("create_object_pipeline", [{"id": "123456789", "file_name": "hat.vtp"}])
-    assert server.compare_image(3, "viewer/create_object_pipeline.jpeg") == True
-
-    server.call("delete_object_pipeline", [{"id": "123456789"}])
-    assert server.compare_image(3, "viewer/delete_object_pipeline.jpeg") == True
-
-
-def test_toggle_object_visibility(server):
-    server.call("create_object_pipeline", [{"id": "123456789", "file_name": "hat.vtp"}])
-    assert server.compare_image(3, "viewer/create_object_pipeline.jpeg") == True
-
-    server.call("toggle_object_visibility", [{"id": "123456789", "visibility": False}])
-    assert server.compare_image(3, "viewer/toggle_object_visibility_1.jpeg") == True
-
-    server.call("toggle_object_visibility", [{"id": "123456789", "visibility": True}])
-    assert server.compare_image(3, "viewer/toggle_object_visibility_2.jpeg") == True
-
-
-def test_apply_textures(server):
-
-    server.call(
-        "create_object_pipeline",
-        [{"id": "123456789", "file_name": "hat.vtp"}],
-    )
-    assert server.compare_image(3, "viewer/create_object_pipeline.jpeg") == True
-
-    server.call(
-        "apply_textures",
-        [
-            {
-                "id": "123456789",
-                "textures": [
-                    {
-                        "texture_name": "lambert2SG",
-                        "texture_file_name": "hat_lambert2SG.vti",
-                    }
-                ],
-            }
-        ],
-    )
-    assert server.compare_image(3, "viewer/apply_textures.jpeg") == True
-
-
 def test_get_point_position(server):
 
     server.call(
-        "create_object_pipeline",
+        "object.register",
         [{"id": "123456789", "file_name": "hat.vtp"}],
     )
-    assert server.compare_image(3, "viewer/create_object_pipeline.jpeg") == True
+    assert server.compare_image(3, "object/register.jpeg") == True
 
     server.call("get_point_position", [{"x": 0, "y": 0}])
     response = server.get_response()
@@ -87,10 +38,10 @@ def test_get_point_position(server):
 def test_take_screenshot(server):
     # Create an object
     server.call(
-        "create_object_pipeline",
+        "object.register",
         [{"id": "123456789", "file_name": "hat.vtp"}],
     )
-    assert server.compare_image(3, "viewer/create_object_pipeline.jpeg") == True
+    assert server.compare_image(3, "object/register.jpeg") == True
 
 
     # Take a screenshot with background jpg
